requirements.txt
auth0-python==3.16.1
Django==3.2
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-gis==0.17
geocoder==1.38.1
Fiona==1.8.19
Shapely==1.7.1
geopandas==0.9.0
msal==1.11.0
numpy==1.20.2
pandas==1.2.4
psycopg2-binary==2.8.6
requests==2.22.0
virtualenv==20.4.4